+++ chronosphere/enum/enum_test.go
@@ -66,10 +66,12 @@ func TestEnumValidateError(t *testing.T) {
 
 func TestAllEnumsValidate(t *testing.T) {
 	v1Spec := loadSwaggerSpec(t, "../pkg/configv1/swagger.json")
+	unstableSpec := loadSwaggerSpec(t, "../pkg/configunstable/swagger.json")
 
 	tests := []struct {
 		legacySwaggerName string
 		v1SwaggerName     string
+		swaggerSpec       *openapi2.T
 		enum              Enum[string, string]
 	}{
 		{
@@ -97,10 +99,24 @@ func TestAllEnumsValidate(t *testing.T) {
 			v1SwaggerName:     "ResponderResponderType",
 			enum:              OpsgenieResponderType.ToStrings(),
 		},
+		{
+			swaggerSpec:   unstableSpec,
+			v1SwaggerName: "ResourceAttributesFlattenMode",
+			enum:          ResourceAttributesFlattenMode.ToStrings(),
+		},
+		{
+			swaggerSpec:   unstableSpec,
+			v1SwaggerName: "ResourceAttributesFilterMode",
+			enum:          ResourceAttributesFilterMode.ToStrings(),
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.enum.Name(), func(t *testing.T) {
-			for _, v := range loadEnumValues(t, v1Spec, tt.v1SwaggerName) {
+			swaggerSpec := v1Spec
+			if tt.swaggerSpec != nil {
+				swaggerSpec = tt.swaggerSpec
+			}
+			for _, v := range loadEnumValues(t, swaggerSpec, tt.v1SwaggerName) {
 				require.Nil(t, tt.enum.Validate(v, nil))
 			}
 		})